Adds Components/Buttons namespace
