+++ src/Components/Buttons/index.js
@@ -1,0 +1,7 @@
+import Button from './Button';
+import ConfirmButton from './ConfirmButton';
+
+export {
+  Button,
+  ConfirmButton
+};
