+++ tests/storybook/components/buttons/button.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { storiesOf, action } from '@storybook/react';
-import classNames from 'classnames';
-import Button from 'Components/Button';
-import ConfirmButton from 'Components/ConfirmButton';
 import { withKnobs, boolean } from '@storybook/addon-knobs';
+import classNames from 'classnames';
+import { Button, ConfirmButton } from 'Components/Buttons';
 /* eslint-disable react/prop-types*/
 
 storiesOf('Buttons', module)
