+++ tests/storybook/components/buttons/icon.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, boolean } from '@storybook/addon-knobs';
 import classNames from 'classnames';
-import Button from 'Components/Button';
+import { Button } from 'Components/Buttons';
 import Icon from 'Components/Icon';
 
 /* eslint-disable react/prop-types*/
