+++ tests/storybook/components/buttons/square.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import Button from 'Components/Button';
+import { Button } from 'Components/Buttons';
 import Icon from 'Components/Icon';
 
 storiesOf('Buttons', module)
